util성 simple text input 패키지 분리
